Create literally all of the site
Wow, I had a wild time getting the graphic design post to work. For seemingly no reason, it wouldn't show up. I used a different file with the same name and content in the end and that ended up working. In that same document I had to use some HTML because i couldn't properly centre align or size the photos without it.